+++ .env.example
@@ -1,6 +1,10 @@
 # Database
 DATABASE_URL=file:./dev.sqlite
 
-# Github Auth Provider
+# Github OAuth Provider
 GITHUB_ID=
 GITHUB_SECRET=
+
+# Google OAuth Provider
+GOOGLE_ID=
+GOOGLE_SECRET=
