+++ src/app/page.tsx
@@ -1,15 +1,8 @@
-'use client'
-import { signIn } from 'next-auth/react'
 import Image from 'next/image'
-import Link from 'next/link'
 
 import logo from '@/assets/logo.svg'
-import { ROUTES } from '@/constants'
-import { Github, Google, RocketLaunch } from '@/icons'
 
-async function handleSignInWithGithub() {
-  await signIn('github')
-}
+import { SignInButtons } from './SignInButtons'
 
 export default function Home() {
   return (
@@ -34,28 +27,7 @@ export default function Home() {
             </p>
           </div>
 
-          <div className="w-full flex flex-col gap-4">
-            <button className="w-full flex items-center text-button-lg font-bold bg-gray-600 py-5 px-6 gap-5 rounded-lg transition-colors hover:bg-gray-500 active:bg-gray-700">
-              <Google />
-              Entrar com Google
-            </button>
-
-            <button
-              className="w-full flex items-center text-button-lg font-bold bg-gray-600 py-5 px-6 gap-5 rounded-lg transition-colors hover:bg-gray-500 active:bg-gray-700"
-              onClick={handleSignInWithGithub}
-            >
-              <Github />
-              Entrar com Github
-            </button>
-
-            <Link
-              href={ROUTES.dashboard}
-              className="w-full flex items-center text-button-lg font-bold bg-gray-600 py-5 px-6 gap-5 rounded-lg transition-colors hover:bg-gray-500 active:bg-gray-700"
-            >
-              <RocketLaunch size={32} className="text-purple-100" />
-              Acessar como visitante
-            </Link>
-          </div>
+          <SignInButtons />
         </div>
       </main>
     </div>
